refactor(phonemes-result): [11] seperate into multiple files
